requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asyncer==0.0.7
attrs==25.1.0
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chainlit==2.1.2
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
colorama==0.4.6
Crawl4AI==0.4.248
cryptography==44.0.0
cssselect==1.2.0
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
duckduckgo_search==7.3.1
fake-http-header==0.3.5
fake-useragent==2.0.3
fastapi==0.115.8
filelock==3.17.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.0.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0rc1
greenlet==3.1.1
groq==0.18.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Lazify==0.4.0
litellm==1.60.6
literalai==0.1.103
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.25.2
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.2
openai==1.61.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
playwright==1.50.0
primp==0.12.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.18
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
starlette==0.41.3
streamlit==1.42.0
syncer==2.0.3
tenacity==9.0.0
tf-playwright-stealth==1.1.1
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uptrace==1.29.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==0.20.0
websockets==14.2
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
